rename bmc due to confusion as the bmc from r2 is different
