+++ keyboards/exclusive/e6v2/oe_bmc/config.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define PRODUCT_ID      0x0000
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    exclusive
-#define PRODUCT         e6v2 bmc
+#define PRODUCT         e6v2 oe bmc
 #define DESCRIPTION     A custom 60% keyboard
 
 /* key matrix size */
