+++ keyboards/exclusive/e6v2/oe_bmc/info.json
