+++ keyboards/exclusive/e6v2/oe_bmc/keymaps/default/readme.md
