+++ keyboards/exclusive/e6v2/oe_bmc/oe_bmc.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "bmc.h"
+#include "oe_bmc.h"
 #include "rgblight.h"
 #include "i2c_master.h"
